+++ src/syntax/syntax.grammar
@@ -1,4 +1,5 @@
-@top Program { FrontMatterSplitter? (expression | Import | Binding)* }
+
+@top Program { FrontMatter? FrontMatterSplitter? (expression | Import | Binding)* }
 
 @skip { space | BlockComment }
 
@@ -26,6 +27,10 @@ OpcodeOpening {
     blockCommentEnd { "*/" }
     blockCommentNewline { "\n" | "\r" | "\r\n" }
     @else blockCommentContent
+}
+
+@external tokens frontMatter from "./tokens" {
+    FrontMatter
 }
 
 @skip {} {
@@ -88,6 +93,8 @@ OpcodeOpening {
    
 }
 
-@external propSource rainHighlight from "./highlight"
+@external propSource highlight from "./props"
+@external propSource indent from "./props"
+@external propSource fold from "./props"
 
 @detectDelim
